Add Tile types
